locust-image/requirements.txt
locust==1.2.3
requests
google-auth
pyjwt
cryptography
